WW-4514 Extracts parameters string building into a dedicated class with a proper extension point
