+++ core/src/main/resources/org/apache/struts2/default.properties
@@ -279,6 +279,12 @@ struts.ognl.expressionMaxLength=256
 ### These formatters are using a slightly different patterns, please check JavaDocs of both and more details is in WW-5016
 struts.date.formatter=dateTimeFormatter
 
+### Defines which instance of ParametersStringBuilder to use, Struts provides just one instance:
+### - strutsParametersStringBuilder
+### The builder is used by UrlHelp to create a proper query string out of provided parameters map
+struts.url.parametersStringBuilder=strutsParametersStringBuilder
+
+### Defines which instances of encoder and decoder to use, Struts provides one default implementation for each
 struts.url.encoder=strutsUrlEncoder
 struts.url.decoder=strutsUrlDecoder
