+++ core/src/main/resources/struts-default.xml
@@ -313,6 +313,8 @@
     <bean type="com.opensymphony.xwork2.ognl.BeanInfoCacheFactory" name="struts"
           class="com.opensymphony.xwork2.ognl.DefaultOgnlBeanInfoCacheFactory" scope="singleton"/>
 
+    <bean type="org.apache.struts2.url.ParametersStringBuilder" name="strutsParametersStringBuilder"
+          class="org.apache.struts2.url.StrutsParametersStringBuilder" scope="singleton"/>
     <bean type="org.apache.struts2.url.UrlEncoder" name="strutsUrlEncoder"
           class="org.apache.struts2.url.StrutsUrlEncoder" scope="singleton"/>
     <bean type="org.apache.struts2.url.UrlDecoder" name="strutsUrlDecoder"
